# openapi.yaml
openapi: 3.0.3
info:
  title: Smarthouse Controller API
  description: |
    REST API for the SmartHouse home automation backend.
    Base URL: `http://192.168.0.201:24867/smarthouse`
  version: 1.0.0
servers:
  - url: http://192.168.0.201:24867/smarthouse
    description: Local server
...
paths:
...
  /interview/topics:
    get:
      tags: [Interview]
      summary: Get topic scoreboard
      description: Returns the current state of topics (scoreboard) including average scores, total/attempted questions, and status.
      operationId: getInterviewTopics
      parameters:
        - name: section
          in: query
          required: false
          description: Filter topics by section (e.g. '14 ORM & Hibernate')
          schema:
            type: string
      responses:
        "200":
          description: List of topics with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewTopicScore"
    post:
      tags: [Interview]
      summary: Create a new topic
      description: Creates a new interview topic.
      operationId: createInterviewTopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewTopicRequest"
      responses:
        "201":
          description: Created topic ID
          content:
            text/plain:
              schema:
                type: integer
  /interview/topics/{id}:
    patch:
      tags: [Interview]
      summary: Update an existing topic
      description: Partially updates an interview topic.
      operationId: updateInterviewTopic
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the topic
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewTopicRequest"
      responses:
        "200":
          description: Topic updated successfully
  /interview/questions:
    get:
      tags: [Interview]
      summary: List interview questions
      description: Returns all questions. Optionally filter by section or topic. Includes the user's latest score and notes for each question.
      operationId: getInterviewQuestions
      parameters:
        - name: section
          in: query
          required: false
          description: Filter questions by section (e.g. '14 ORM & Hibernate')
          schema:
            type: string
        - name: topicId
          in: query
          required: false
          description: Filter questions by a specific topic ID
          schema:
            type: integer
      responses:
        "200":
          description: List of questions with their latest scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewQuestion"
    post:
      tags: [Interview]
      summary: Create a new question
      description: Creates a new interview question.
      operationId: createInterviewQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewQuestionRequest"
      responses:
        "201":
          description: Created question ID
          content:
            text/plain:
              schema:
                type: integer
  /interview/questions/{id}:
    patch:
      tags: [Interview]
      summary: Update a question
      description: Partially updates an existing interview question.
      operationId: updateInterviewQuestion
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the question
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewQuestionRequest"
      responses:
        "200":
          description: Question updated successfully
  /interview/questions/next:
    get:
      tags: [Interview]
      summary: Get the next interview question
      description: Returns the next question to ask based on priority (Not Started topics first, then lowest score). Ensures no consecutive repeated questions.
      operationId: getNextInterviewQuestion
      parameters:
        - name: section
          in: query
          required: true
          description: Section to pull the question from
          schema:
            type: string
      responses:
        "200":
          description: The next interview question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewQuestion"
        "404":
          description: No more questions available for this section
  /interview/questions/{id}/attempts:
    get:
      tags: [Interview]
      summary: Get all attempts for a question
      description: Returns the history of attempts for a specific question, ordered by most recent first.
      operationId: getInterviewAttempts
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the question
          schema:
            type: integer
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewAttempt"
    post:
      tags: [Interview]
      summary: Submit a question attempt
      description: Records a score and notes for a question attempt, and returns the updated topic score.
      operationId: submitInterviewAttempt
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the question
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewAttemptRequest"
      responses:
        "200":
          description: Updated topic score after the attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewTopicScore"

  # ── Appliances ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InterviewTopicScore:
      type: object
      properties:
        id:
          type: integer
        topicName:
          type: string
        totalQuestions:
          type: integer
        attemptedQuestions:
          type: integer
        avgScore:
          type: number
          format: double
          nullable: true
        status:
          type: string
        jdRelevance:
          type: integer
          nullable: true
    InterviewQuestion:
      type: object
      properties:
        id:
          type: integer
        questionText:
          type: string
        difficulty:
          type: string
        topicName:
          type: string
        topicId:
          type: integer
        score:
          type: integer
          nullable: true
          description: The latest attempt score (if any)
        notes:
          type: string
          nullable: true
          description: Notes from the latest attempt
    InterviewAttempt:
      type: object
      properties:
        id:
          type: integer
        score:
          type: integer
          description: Score from 0 to 100
        notes:
          type: string
          description: Feedback notes on the answer
    InterviewAttemptRequest:
      type: object
      required: [score, notes]
      properties:
        score:
          type: integer
          description: Score from 0 to 100
        notes:
          type: string
          description: Feedback notes on the answer
    InterviewTopicRequest:
      type: object
      properties:
        section:
          type: string
          description: e.g. '14 ORM & Hibernate'
        topicName:
          type: string
        jdRelevance:
          type: integer
        filePath:
          type: string
          description: Path to the study notes markdown file
    InterviewQuestionRequest:
      type: object
      properties:
        topicId:
          type: integer
        questionText:
          type: string
        difficulty:
          type: string